virtu_fit_ia_v2/requirements.txt
fastapi 
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
optuna
psycopg2-binary
python-dotenv
xgboost
plotly